Added EazyWeld Plugin and UI Demo
EazyWeld is a tool I wrote to assist in the construction of complicated models. It forms "welds" between a selected (or clicked on) "base part" and any other part inside a model.

These welds allow the model to move and interact within the physics environment as a unit, rather than falling apart when a single part is interacted with.

A demo is also attached of how the plugin appears in the workspace editor.